libcamera: Rebased the Raspberry pi pipeline to master
Rebased the kbingham Raspberry pi pieline changes to fit
to the latest libcamera master.

Removed the Kbingham isp changes in this pipeline, Need to
add in future.

With this cam utility works fine, with cam utility Raspberry pi
camera get detected and it could capture and store the frames.
But the frames are not raw image.

Signed-off-by: Madhavan Krishnan <madhavan.krishnan@linaro.org>